+++ home-manager/modules/programs/git.nix
@@ -45,7 +45,7 @@ in
       };
 
       aliases = {
-        fixup = "!git log -n 50 --pretty=format:'%h %s' --no-merges | fzf | cut -c -7 | xargs -o git commit --fixup";
+        fixup = mkDefault "!git log -n 50 --pretty=format:'%h %s' --no-merges | fzf | cut -c -7 | xargs -o git commit --fixup";
       };
 
       extraConfig = {