+++ nixos/modules/profile/usage/workstation.nix
@@ -11,7 +11,7 @@
 { config, lib, pkgs, ... }:
 
 let
-  inherit (lib) mkIf;
+  inherit (lib) mkDefault mkIf;
 
   hostName = config.networking.hostName;
   layout = config.confkit.keyboard.layout;
@@ -75,13 +75,13 @@ in
     ########################################################################
 
     hardware = {
-      pulseaudio.enable = true;
+      pulseaudio.enable = mkDefault true;
     };
 
     sound = {
       # Enable ALSA sound.
-      enable = true;
-      mediaKeys.enable = true;
+      enable = mkDefault true;
+      mediaKeys.enable = mkDefault true;
     };
 
     ########################################################################
@@ -89,7 +89,7 @@ in
     ########################################################################
 
     networking = {
-      networkmanager.enable = true;
+      networkmanager.enable = mkDefault true;
     };
 
     ########################################################################
@@ -97,12 +97,12 @@ in
     ########################################################################
 
     services = {
-      pcscd.enable = true;
-      printing.enable = true;
-      udisks2.enable = true;
+      pcscd.enable = mkDefault true;
+      printing.enable = mkDefault true;
+      udisks2.enable = mkDefault true;
 
       xserver = {
-        enable = true;
+        enable = mkDefault true;
 
         # Configure the keyboard layout if it has been set in
         # confkit.keyboard.layout.
@@ -111,8 +111,8 @@ in
 
         # Enable touchpad support with natural scrolling.
         libinput = {
-          enable = true;
-          touchpad.naturalScrolling = true;
+          enable = mkDefault true;
+          touchpad.naturalScrolling = mkDefault true;
         };
       };
     };